Renamed module to match other adapters.
