+++ activerecord-postgis-adapter.gemspec
@@ -1,10 +1,10 @@
-require './lib/active_record/connection_adapters/postgis_adapter/version.rb'
+require './lib/active_record/connection_adapters/postgis/version.rb'
 
 ::Gem::Specification.new do |spec|
   spec.name = 'activerecord-postgis-adapter'
   spec.summary = 'ActiveRecord adapter for PostGIS, based on RGeo.'
   spec.description = "ActiveRecord connection adapter for PostGIS. It is based on the stock PostgreSQL adapter, but provides built-in support for the spatial extensions provided by PostGIS. It uses the RGeo library to represent spatial data in Ruby."
-  spec.version = ActiveRecord::ConnectionAdapters::PostGISAdapter::VERSION
+  spec.version = ActiveRecord::ConnectionAdapters::PostGIS::VERSION
   spec.author = 'Daniel Azuma, Tee Parham'
   spec.email = 'dazuma@gmail.com, parhameter@gmail.com'
   spec.homepage = "http://github.com/rgeo/activerecord-postgis-adapter"
@@ -21,5 +21,4 @@ require './lib/active_record/connection_adapters/postgis_adapter/version.rb'
 
   spec.add_development_dependency 'rake', '~> 10.2'
   spec.add_development_dependency 'mocha', '~> 1.0'
-  spec.add_development_dependency 'appraisal', '~> 1.0'
 end
