+++ lib/active_record/connection_adapters/postgis/create_connection.rb
@@ -8,7 +8,7 @@ module ActiveRecord  # :nodoc:
   module ConnectionHandling  # :nodoc:
     if(defined?(::RUBY_ENGINE) && ::RUBY_ENGINE == 'jruby')
       def postgis_connection(config)
-        config[:adapter_class] = ::ActiveRecord::ConnectionAdapters::PostGISAdapter::MainAdapter
+        config[:adapter_class] = ::ActiveRecord::ConnectionAdapters::PostGIS::MainAdapter
         postgresql_connection(config)
       end
       alias_method :jdbcpostgis_connection, :postgis_connection
@@ -30,7 +30,7 @@ module ActiveRecord  # :nodoc:
 
         # The postgres drivers don't allow the creation of an unconnected PGconn object,
         # so just pass a nil connection object for the time being.
-        ::ActiveRecord::ConnectionAdapters::PostGISAdapter::MainAdapter.new(nil, logger, conn_params, config)
+        ::ActiveRecord::ConnectionAdapters::PostGIS::MainAdapter.new(nil, logger, conn_params, config)
       end
     end
