+++ lib/active_record/connection_adapters/postgis/databases.rake
@@ -8,7 +8,7 @@ namespace :db do
       environments_ << 'test' if ::Rails.env.development?
       configs_ = ::ActiveRecord::Base.configurations.values_at(*environments_).compact.reject{ |config_| config_['database'].blank? }
       configs_.each do |config_|
-        ::ActiveRecord::ConnectionAdapters::PostGISAdapter::PostGISDatabaseTasks.new(config_).setup_gis
+        ::ActiveRecord::ConnectionAdapters::PostGIS::PostGISDatabaseTasks.new(config_).setup_gis
       end
     end
   end
