+++ lib/active_record/connection_adapters/postgis/postgis_database_tasks.rb
