+++ lib/active_record/connection_adapters/postgis/railtie.rb
@@ -3,7 +3,7 @@ require 'active_record/connection_adapters/postgis_adapter'
 
 module ActiveRecord  # :nodoc:
   module ConnectionAdapters  # :nodoc:
-    module PostGISAdapter  # :nodoc:
+    module PostGIS  # :nodoc:
       class Railtie < ::Rails::Railtie  # :nodoc:
         rake_tasks do
           load ::File.expand_path("databases.rake", ::File.dirname(__FILE__))
