+++ lib/active_record/connection_adapters/postgis/spatial_column.rb
@@ -1,6 +1,6 @@
 module ActiveRecord  # :nodoc:
   module ConnectionAdapters  # :nodoc:
-    module PostGISAdapter  # :nodoc:
+    module PostGIS  # :nodoc:
       class SpatialColumn < ConnectionAdapters::PostgreSQLColumn  # :nodoc:
 
         def initialize(factory_settings, table_name, name, default, oid_type, sql_type = nil, null = true, opts = nil)
