+++ lib/active_record/connection_adapters/postgis/version.rb
@@ -1,7 +1,7 @@
 module ActiveRecord
   module ConnectionAdapters
-    module PostGISAdapter
-      VERSION = '2.2.0'.freeze
+    module PostGIS
+      VERSION = '3.0.0'.freeze
     end
   end
 end
