+++ lib/active_record/connection_adapters/postgis_adapter.rb
@@ -3,7 +3,7 @@
 
 module ActiveRecord
   module ConnectionAdapters
-    module PostGISAdapter
+    module PostGIS
       # The name returned by #adapter_name
       ADAPTER_NAME = 'PostGIS'.freeze
     end
@@ -16,21 +16,21 @@ require 'active_record'
 require 'active_record/connection_adapters/postgresql_adapter'
 require 'rgeo/active_record'
 
-require 'active_record/connection_adapters/postgis_adapter/version'
-require 'active_record/connection_adapters/postgis_adapter/common_adapter_methods'
-require 'active_record/connection_adapters/postgis_adapter/main_adapter'
-require 'active_record/connection_adapters/postgis_adapter/spatial_column_info'
-require 'active_record/connection_adapters/postgis_adapter/spatial_table_definition'
-require 'active_record/connection_adapters/postgis_adapter/spatial_column'
-require 'active_record/connection_adapters/postgis_adapter/arel_tosql'
-require 'active_record/connection_adapters/postgis_adapter/setup'
-require 'active_record/connection_adapters/postgis_adapter/create_connection'
-require 'active_record/connection_adapters/postgis_adapter/postgis_database_tasks'
+require 'active_record/connection_adapters/postgis/version'
+require 'active_record/connection_adapters/postgis/common_adapter_methods'
+require 'active_record/connection_adapters/postgis/main_adapter'
+require 'active_record/connection_adapters/postgis/spatial_column_info'
+require 'active_record/connection_adapters/postgis/spatial_table_definition'
+require 'active_record/connection_adapters/postgis/spatial_column'
+require 'arel/visitors/postgis'
+require 'active_record/connection_adapters/postgis/setup'
+require 'active_record/connection_adapters/postgis/create_connection'
+require 'active_record/connection_adapters/postgis/postgis_database_tasks'
 
-::ActiveRecord::ConnectionAdapters::PostGISAdapter.initial_setup
+::ActiveRecord::ConnectionAdapters::PostGIS.initial_setup
 
 if defined?(::Rails::Railtie)
-  load ::File.expand_path('postgis_adapter/railtie.rb', ::File.dirname(__FILE__))
+  load ::File.expand_path('postgis/railtie.rb', ::File.dirname(__FILE__))
 end
 
 # :startdoc:
